requirements.txt
llama-index-core~=0.10.20
llama-index-readers-file~=0.1.11
llama-index-embeddings-huggingface~=0.1.4
llama-index-llms-huggingface~=0.1.4
llama-index-retrievers-bm25~=0.1.3
huggingface_hub[inference]~=0.20.3
streamlit~=1.32.2